Fixed conflicting GUIDs, updated PlayerControl
Some assets were moved, and Github either didn't track the .meta file changes, or the deletion of the original files. This caused some unity GUIDs to duplicate, and caused some missing assets.

Also, the new PlayerControl script was not compatible with the Actor class it inherited from. Removed some redundant code that was delegated to Actor some time ago.