Added Ice shard spell.
Adds ice shard projectile. For the moment, is launched with the 'p' key.
Actors damaged by it will be frozen temporarily: This halts AI and movement, and has a visual effect attached.

There is a special type of water tile in the water_dungeon tileset that will freeze into a walkable platform when hit by this ice shard.

Actor code, and existing enemy code, has been changed to acommodate these effects.

Hitbox and projectile code was also tweaked to reflect a change in the way certain actors (skeleton and slime) flip their sprites (now uses negative X scale rather than just flipping the sprite).

Fixed a minor bug in the Navigator that cused it to complain if not given information for every blocking type.