Establishes "Saves" directory. Player data saves.
Player data can now be successfully set when loading a save.
Saved games (when playing in-editor) are now located in "Assets/Saves/", and .gitIgnore has been
altered to ignore this directory.
For builds, saved games will be located in a data directory next to the
executable, and in a "Saves" subdirectory.