Range of fixes and small additions.
Fixes what should be the last few bugs in the player prefab.
Fireball and push spells can be saved and loaded from/to the inventory.
Coins are now dropped by enemies, in three different sizes.
Spellcasting and other things now are blocked while the game is paused.
The Dev console has been fixed, and has gained the following commands:
	- saveinv: used to save the inventory to the live save.
	- loadinv: used to load the inventory from the live save.